backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.2
Django==3.2.3
djangorestframework==3.12.4
drf-yasg==1.21.7
exceptiongroup==1.1.3
flower==2.0.1
gunicorn==20.1.0
humanize==4.8.0
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.2
mypy==1.6.0
mypy-extensions==1.0.0
packaging==23.2
pluggy==1.3.0
prometheus-client==0.8.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.3
pytest==7.4.2
pytest-celery==0.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tornado==6.3.3
transliterate==1.10.2
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.8
